pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vidnavigator"
version = "1.0.0"
authors = [
  { name="Hatem", email="hatem.mezlini@vidnavigator.com" },
]
description = "Python SDK for the VidNavigator Developer API"
readme = "README.md"
requires-python = ">=3.7"
license = { file="LICENSE" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "requests>=2.31.0",
    "pydantic>=1.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "python-dotenv>=0.19",
]

[project.urls]
"Homepage" = "https://vidnavigator.com"
"Repository" = "https://github.com/vidnavigator/vidnavigator-python"
"Bug Tracker" = "https://github.com/vidnavigator/vidnavigator-python/issues"

[tool.setuptools.packages.find]
exclude = ["tests*"] 